# Remove file reading for FCI L1c