--- a/satpy/etc/readers/fci_l1c_nc.yaml
+++ b/satpy/etc/readers/fci_l1c_nc.yaml
@@ -7,7 +7,7 @@ reader:
     Used to read Meteosat Third Generation (MTG) Flexible
     Combined Imager (FCI) L1c data.
   status: Beta for full-disc FDHSI and HRFI, RSS not supported yet
-  supports_fsspec: false
+  supports_fsspec: true
   reader: !!python/name:satpy.readers.yaml_reader.GEOVariableSegmentYAMLReader
   sensors: [ fci ]