@pnuu 2022-12-02
Initial timing tests using local files and test script from https://github.com/pytroll/satpy/pull/2237#issuecomment-1282332336:

main:
```
Done. It took 3.73s  to start Scene.
Done. It took 2.63s to load datasets.
Done. It took 11.24s  to save datasets.
Done. It took 17.59s  for all ops.
```

this PR:
```
Done. It took 27.46s  to start Scene.
Done. It took 5.02s  to load datasets.
Done. It took 11.81s  to save datasets.
Done. It took 44.29s  for all ops.
```

I'll run some tests with data on S3 next week.