@ameraner 2022-12-02
Do you have a plan for switching (automagically?) between `NetCDF4FsspecFileHandler` and `NetCDF4FileHandler`, depending if the data is remote or not, to avoid the performance penalty with local data?