@pnuu 2022-12-02
This is one way I found for the automatic backend selection: https://github.com/pnuu/satpy/blob/feature_optimise_fcil1c_only_necessary_h5/satpy/readers/netcdf_utils.py#L363