@pnuu 2022-12-04
I now transferred the code I linked to this PR. Will need unit tests and testing with real data. I hope to have time for those on Wednesday. I might have a look at the currently failing tests even before that.