@pnuu 2022-12-05
The file handler caches were not cleared, so the values from previous tests were re-used in the later tests. Well, at least the caching works :joy: