@pnuu 2022-12-07
Tested with data on internal CEPH/S3, and the processing works. Due to network delay, the processing time isn't great. Looking at the graph, the downloading part was 72 s of 94 s full processing time. I'll see if I can test again on a server closer to the bucket. Here I just created single `airmass` composite and resampled to 500 m EPSG:3035 grid (7836 x 7736 pixels).

![Screenshot 2022-12-07 at 16-34-23 Bokeh Plot](https://user-images.githubusercontent.com/3170788/206207223-07539da0-f33b-4d2f-a358-553864df2513.png)