@pnuu 2022-12-07
Damn, tested the updated code with local files, and it fails in platform name handling. I'll have a look tomorrow (I hope).