@pnuu 2022-12-07
Ok, the problem is that `self['attr/platform']` is used to get the translation from short platform name to full platform name, and the the value of it is this dictionary:

```python
ipdb> print(self['attr/platform'])
{'Conventions': 'CF-1.7', 'title': 'FCI Level 1C Rectified Radiance Product', 'summary': 'Body chunk of the MTG FCI Level 1c dataset for the FDHSI subset of spectral channels', 'keywords': 'MTG FCI FDHSI level_1c', 'keywords_vocabulary': 'original generated file', 'history': 'Original generated version', 'institution': 'EUMETSAT', 'location_indicator': 'XX-EUMETSAT-Darmstadt', 'data_designator': 'IMG+SAT', 'platform': 'MTI1', 'data_source': 'FCI', 'processing_level': '1C', 'type': 'RRAD', 'subtype': 'FDHSI', 'component1': 'CHK', 'component3': 'null', 'product_id': 'FCI1CRRADFDHSI', 'baseline_version': 'n.x.y', 'release_version': 'original', 'processor_version': '<configured_value>', 'algorithm_version': '<configured_value>', 'processing_mode': 'N', 'subsetting': 'null', 'source': 'IDPFI', 'runtime_data': '<runtime_value>', 'parent_data': '<runtime_value>', 'linked_data': '<runtime_value>', 'references': 'www.eumetsat.int', 'comment': 'None', 'group_tag': 'MTI1_FCI_1C_RRAD_FDHSI_YYY_DDD_NNNN_original', 'instrument_configuration_id': '<runtime_value>', 'instrument_configuration_id_version': '<runtime_value>', 'subsettable_groups': 'vis_04 vis_05 vis_06 vis_08 vis_09 nir_13 nir_16 nir_22 ir_38 wv_63 wv_73 ir_87 ir_97 ir_105 ir_123 ir_133', 'subsettable_groups_present': 'vis_04 vis_05 vis_06 vis_08 vis_09 nir_13 nir_16 nir_22 ir_38 wv_63 wv_73 ir_87 ir_97 ir_105 ir_123 ir_133', 'mtg_name': '<runtime_value>', 'alternative_name': 'null', 'format': 'NC4E', 'id': '<configured_value>', 'naming_authority': '<configured_value>', 'creator_type': 'institution', 'creator_institution': 'EUMETSAT', 'creator_url': 'www.eumetsat.int', 'license': 'null', 'standard_name_vocabulary': 'CF Standard Name Table v27', 'project': 'MTG', 'time_coverage_duration': 'null', 'time_coverage_resolution': 'null', 'cdm_datatype': 'null', 'coverage': 'FD', 'component2': 'BODY', 'format_version': '4c', 'time_coverage_start': '20170920120008', 'time_coverage_end': '20170920120015', 'disposition_mode': 'T', 'facility_or_tool': 'GTT', 'environment': 'DEV', 'date_created': '20170920120515', 'repeat_cycle_in_day': '0073', 'processed_count_in_repeat_cycle': '0001', 'count_in_repeat_cycle': '0001', 'purpose': array([], dtype=float64), 'date_time_position': '20170920120000', 'time_position': '120000', 'geospatial_lat_min': -81.19571706772024, 'geospatial_lat_max': -69.26566506254072, 'geospatial_lon_min': -52.67815684599577, 'geospatial_lon_max': 52.67815684599568, 'creator_name': 'EUMETSAT', 'creator_email': 'ops@eumetsat.int', 'special_compression': 'JLS'}
```

I'll try to hunt the origin of this change. Tomorrow. Hopefully.