@pnuu 2022-12-08
It was a stupid mistake, using `super()._get_object_attrs(obj)` instead of `super()._get_attr(obj, key)` when not using `h5netcdf` backend.