@pnuu 2022-12-08
I don't think I can add more meaningful tests without replacing all the fake file handlers with actual files that are written by the tests. And that is such a big task that I'm not willing to do that within this PR.