@pnuu 2022-12-08
Here are the tests running on a server. Only two cores, `airmass` resampled to the 7836 x 7736 pixel area.

Using files NAS mounted as local file system (so using `netCDF4` backend):
![Screenshot 2022-12-08 at 12-35-44 Bokeh Plot](https://user-images.githubusercontent.com/3170788/206425260-8d4aacab-2dba-4951-880e-cc1f37fba8be.png)

Reading the files from CEPH/S3 (same network speed as for the NAS):
![Screenshot 2022-12-08 at 12-39-24 Bokeh Plot](https://user-images.githubusercontent.com/3170788/206426919-a3a72067-5014-4088-862b-5dc50a5b0fb8.png)

The bit until 100 seconds includes the file transfer and (slow) file handler creation. Using `h5netcdf` backend for local files (commenting out the try/except in `NetCDF4FsspecFileHandler._get_file_handle()`) shows that the CEPH/S3 reading doesn't slow down the reading/file handling at all:
![Screenshot 2022-12-08 at 12-57-36 Bokeh Plot](https://user-images.githubusercontent.com/3170788/206429774-c49f3937-609d-4ba9-a792-09bd25f5146d.png)