@pnuu 2023-01-10
After merging `main` also the unstable tests now pass.

The remaining question is [this](https://github.com/pytroll/satpy/pull/2305#issuecomment-1342297927): should the tests write a file that mimics the relevant parts of the file? If yes, that should be done in a separate PR as it's a rather large task.