@mraspaud 2023-01-10
> should the tests write a file that mimics the relevant parts of the file?

Imo, for the tests to be comprehensive while still readable, nothing beats having a stub file on temp storage.
Agreed that this should be in a different PR though.