@pnuu 2023-01-13
The lower test coverage is the only remaining bit I think, and as discussed above, that requires another PR where these tests are converted to use a real file with all the required stuff in it.